Move refineUsingParent to TypeOps
