+++ compiler/src/dotty/tools/dotc/transform/TypeUtils.scala
@@ -1,4 +1,5 @@
-package dotty.tools.dotc
+package dotty.tools
+package dotc
 package transform
 
 import core._
